Removed need for Clear-Host
Previously, a long script block was bound to a hotkey, and a call to cls was made to get that spam out of the console.  Now the hotkey is bound to a script block which just calls one command, removing the need for the cls, so you don't lose whatever previous output happened to be in the console.